BreezeBnb/pyproject.toml
[project]
name = "breezebnb"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pyspark (==4.0.1)",
    "findspark (==2.0.1)",
    "python-dotenv (==1.0.1)",
    "pytest (==7.0.0)",
    "chispa (==0.10.0)",
    "pytest-mock (==3.14.1)",
    "pyyaml (==6.0.1)"
]

[tool.poetry]
packages = [{include = "breezebnb", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
